+++ R/get_MIdat.R
@@ -6,13 +6,14 @@
 #' the \href{https://CRAN.R-project.org/package=foreign}{foreign} package needs to be installed.
 #' @param object object inheriting from class \code{JointAI}
 #' @param m number of imputed datasets
+#' @param include should the original, incomplete data be included?
 #' @param start the first iteration of interest (see \code{\link[coda]{window.mcmc}})
 #' @param seed optional seed
+#' @param export_to_SPSS logical; should the completed data be exported to SPSS?
 #' @param resdir optional directory for results (if unspecified and
 #'               \code{export_to_SPSS = TRUE} the current working directory is used)
 #' @param filename optional file name (without ending; if unspecified and
 #'                 \code{export_to_SPSS = TRUE} a name is generated automatically)
-#' @param export_to_SPSS logical
 #'
 #' @return A dataframe containing the imputed values (and original data) stacked.
 #'        The variable \code{Imputation_} identifies the imputations.
@@ -31,8 +32,10 @@
 #' }
 #' @export
 #'
-get_MIdat <- function(object, m = 10, start = NULL, seed = NULL, resdir = NULL,
-                      filename = NULL, export_to_SPSS = FALSE){
+get_MIdat <- function(object, m = 10, include = TRUE,
+                      start = NULL, seed = NULL,
+                      export_to_SPSS = FALSE,
+                      resdir = NULL, filename = NULL){
 
   if (is.null(object$meth))
     stop("This JointAI object did not impute any values.")
@@ -53,6 +56,8 @@ get_MIdat <- function(object, m = 10, start = NULL, seed = NULL, resdir = NULL,
     }
     tvar <- apply(DFlong, 2, check_tvar, groups)
     DF <- DFlong[match(unique(DFlong[, id]), DFlong[, id]), names(tvar)[!tvar]]
+  } else {
+    DF$.id <- 1:nrow(DF)
   }
 
   meth <- object$meth
@@ -159,6 +164,9 @@ get_MIdat <- function(object, m = 10, start = NULL, seed = NULL, resdir = NULL,
     }
   }
 
+
+  if (!include)
+    DF_list <- DF_list[-1]
 
 # build dataset --------------------------------------------------------------------------
   if (object$analysis_type == "lme") {